+++ core/src/main/kotlin/com/falsepattern/zigbrains/project/module/ZigModuleBuilder.kt
@@ -29,6 +29,7 @@ import com.intellij.ide.util.projectWizard.ModuleBuilder
 import com.intellij.ide.util.projectWizard.ModuleWizardStep
 import com.intellij.ide.util.projectWizard.WizardContext
 import com.intellij.openapi.Disposable
+import com.intellij.openapi.application.ModalityState
 import com.intellij.openapi.module.ModuleType
 import com.intellij.openapi.roots.ModifiableRootModel
 import com.intellij.openapi.util.Disposer
@@ -58,7 +59,7 @@ class ZigModuleBuilder: ModuleBuilder() {
         val root = contentEntry.file ?: return
         val config = configurationData ?: return
         config.generateProject(this, rootModel.project, root, forceGitignore)
-        withEDTContext {
+        withEDTContext(ModalityState.defaultModalityState()) {
             root.refresh(false, true)
         }
     }